Able to get a valid JWT for a given user!

+++ souschef-core/Model/DTO/AuthResponse.cs
@@ -3,9 +3,9 @@ namespace souschef_core.Model.DTO;
 /// <summary>
 /// DTO for responding to user authentication requests
 /// </summary>
-public class AuthResponse(int id, string username, string token)
+public class AuthResponse(long id, string username, string token)
 {
-    public int Id { get; init; } = id;
+    public long Id { get; init; } = id;
     public string Username { get; init; } = username;
     public string Token { get; init; } = token;
 }
